% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticPopulation.R
\name{geneticPopulation}
\alias{geneticPopulation}
\title{Evolve a Population}
\usage{
geneticPopulation(fitfunc, species, gens, indivs, creation = "new",
  plot = "maxVals")
}
\arguments{
\item{fitfunc}{A function; used to evaluate the fitness of each individual each generation.}

\item{species}{A scalar; the number of species to create and evolve.}

\item{gens}{A scalar; the number of generations each species should be eveolved over.}

\item{indivs}{A scalar; the number of individuals per species to create and evolve.}

\item{creation}{A character; the}
}
\description{
Evolve a Population
}
