% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticSpecies.R
\name{geneticSpecies}
\alias{geneticSpecies}
\title{Creates and analyzes a genetic species}
\usage{
geneticSpecies(fitfunc, indivs, gens, creation = "new", return = F)
}
\arguments{
\item{fitfunc}{A number.}

\item{indivs}{Another number.}
}
\description{
Creates and analyzes a genetic species
}
